Clean up some code, add some tests
This commit creates the Future concept, changes the interface to the
Reindex method to make more sense, and switches from flags to options for
the Clean method.  It also adds tests for much of the library, though the
coverage is still quite low (only about 35%), and tunes up many of the
doc comments.